pyImageTranslator/requirements.txt
paddlepaddle==2.0.0
Shapely==1.7.1
opencv_python==4.2.0.32
requests==2.21.0
six==1.10.0
lmdb==1.1.1
imgaug==0.4.0
numpy==1.18.4
scipy==1.2.1
scikit_image==0.17.2
pywin32==223
pyclipper==1.2.1
Js2Py==0.71
tqdm==4.31.1
lanms==1.0.2
paddle==1.0.2
paddlehub==2.0.4
Pillow==8.2.0
PyQt5==5.15.4
python_Levenshtein==0.12.2
PyYAML==5.4.1
sip==6.0.3
# visualdl==2.1.1
